Add total debt; credit line; percentage

+++ src/pages/editBank/editBank.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { BankService } from '../../service/bank.service';
-import { NavController, NavParams } from 'ionic-angular';
+import { NavController, NavParams, AlertController } from 'ionic-angular';
 
 @Component({
     selector: 'page-editBank',
@@ -11,13 +11,30 @@ export class EditBankPage {
 
     bank: any;
 
-    constructor(public navCtrl: NavController, private navParams: NavParams, private bankService: BankService) {
+    constructor(public navCtrl: NavController, private navParams: NavParams, private bankService: BankService, public alertCtrl: AlertController) {
         this.bank = this.navParams.get('bank');
-        console.log(this.bank);
     }
+    
     removeBank(bank) {
-        this.bankService.removeBank(bank);
-        this.navCtrl.pop()
+        let confirm = this.alertCtrl.create({
+            title: 'Sure to DELETE?',
+            message: 'Do you want to delete this payment?',
+            buttons: [
+              {
+                text: 'Cancel',
+                handler: () => {
+                }
+              },
+              {
+                text: 'Agree',
+                handler: () => {
+                    this.bankService.removeBank(bank);
+                    this.navCtrl.pop()
+                }
+              }
+            ]
+          });
+          confirm.present();
     }
     updateBank(bank) {
         this.bankService.updateBank(bank);
